+++ CHANGELOG.md
@@ -4,7 +4,7 @@ https://github.com/jcorporation/myMPD/
 
 ***
 
-## myMPD v21.0.1 (not yet released)
+## myMPD v21.0.1 (2025-05-27)
 
 This is a small bug fix release.